Refactor BlogProvider imports and move to context directory for improved structure
